requirements.txt
asgiref==3.8.1
binance-connector==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.1
django-environ==0.11.2
djangorestframework==3.15.2
drf-yasg==1.21.7
envtoml==0.1.2
idna==3.7
inflection==0.5.1
packaging==24.1
pep8==1.7.1
pycryptodome==3.20.0
PyJWT==2.9.0
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.1
toml==0.10.2
uritemplate==4.1.1
urllib3==2.2.2
websocket-client==1.8.0
